pyproject.toml
[tool.poetry]
name = "max_talk"
version = "0.1.0"
description = "Simple messager"
authors = ["MaxST <mstolpasov@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.1"
sqlalchemy = "^1.3"
sqlalchemy-utils = "^0.34.1"
passlib = "^1.7"
pyqt5 = "^5.13"
dynaconf = "^2.0"

[tool.poetry.dev-dependencies]
flake8-broken-line = "^0.1.0"
flake8 = "^3.7"
flake8-builtins = "^1.4"
flake8-coding = "^1.3"
flake8-commas = "^2.0"
flake8-comprehensions = "^2.1"
flake8-quotes = "^2.0"
pep8-naming = "^0.8.2"
ipdb = "^0.12.0"
yapf = "^0.28.0"
pytest = "^5.0"
pytest-cov = "^2.7"
python-language-server = {version = "^0.27.0", extras = ["all"]}
pyls-isort = "^0.1.1"
mccabe = "^0.6.1"
pyls-flake8 = {git = "https://github.com/emanspeaks/pyls-flake8.git"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
